+++ src/doc/Documents.build.shared.ps1
@@ -109,7 +109,7 @@ function Add-BuildOpenDocumentTask
 		$null = $taskParams.Add( 'Version', $Version );
 	};
 
-	$taskJobs = $Jobs;
+	$taskJobs = @( 'XSLT-tools' ) + $Jobs;
 	$taskJobs += `
 	{
 		$destFile = $Outputs[0];